Change parameter type of printSettingOPA from int32_t to size_t for type consistency

+++ src/configuration.c
@@ -948,7 +948,7 @@ static void printSettingJSON(void) {
   printf_("json = %s\r\n", config.baseCfg.useJson ? "on" : "off");
 }
 
-static void printSettingOPA(const int32_t ch) {
+static void printSettingOPA(const size_t ch) {
   printf_("opa%d ", (ch + 1));
 
   /* OneWire */
